Refactoring and improving enhancement
