+++ src/main/java/net/alepuzio/buildCSS/App.java
@@ -1,7 +1,7 @@
 package net.alepuzio.buildCSS;
 
-import net.alepuzio.buildCSS.DevelopmentUtil;
-import net.alepuzio.buildCSS.FlowMoreCSS;
+import net.alepuzio.buildCSS.enumeration.EnumMessages;
+import net.alepuzio.buildCSS.logic.FlowMoreCSS;
 
 public class App {
 	
@@ -10,9 +10,9 @@ public class App {
 	 * @param args: parameters from terminal
 	 * */
     public static void main( String[] args ) throws Exception {
-    	DevelopmentUtil.printMsgDebug("Begin");
+    	DevelopmentUtil.printMsgDebug(EnumMessages.BEGIN_PROGRAM.getValue());
     	FlowMoreCSS.factory(args).createsCSSThemesFromDirectory();
-    	DevelopmentUtil.printMsgDebug("Successfull creation of themes");
+    	DevelopmentUtil.printMsgDebug(EnumMessages.END_PROGRAM.getValue());
     }
 
 }
